requirements.txt
appdirs==1.4.3
attrs==16.3.0
cloudpickle==0.2.2
cymem==1.31.2
cytoolz==0.8.2
decorator==4.0.11
dill==0.2.6
ftfy==4.4.2
h5py==2.7.0
html5lib==0.999999999
Keras==2.0.3
murmurhash==0.26.4
nltk==3.2.2
numpy==1.12.1
packaging==16.8
pandas==0.19.2
pathlib==1.0.1
pexpect==4.2.1
pickleshare==0.7.4
#pkg-resources==0.0.0
plac==0.9.6
preshed==0.46.4
prompt-toolkit==1.0.14
protobuf==3.2.0
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pywsd==1.0.4
PyYAML==3.12
regex==2017.4.5
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
semver==2.7.6
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
spacy==0.101.0
tensorflow==1.1.0
tensorflow-gpu==1.1.0
termcolor==1.1.0
Theano==0.9.0
thinc==5.0.8
toolz==0.8.2
tqdm==4.11.2
traitlets==4.3.2
ujson==1.35
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.1
wrapt==1.10.10
